requirements.txt
Django>=4.0
mysqlclient>=2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0

openpyxl>=3.1.0
